Add tests for store/lock.
